Qcom changing tracking
I cannot build for Nexus 4 with the CM tracking, god knows what they have done but I am not the only one with errors. Xylon had to clone some repo's and mess to get their builds to work also. Maybe we can clone this and make it our own repo to manage? That way we can control merges and what might break a build.